// packages/core/primitives/event-dispatch/src/key_code.d.ts
/**
 * @license
 * Copyright Google LLC All Rights Reserved.
 *
 * Use of this source code is governed by an MIT-style license that can be
 * found in the LICENSE file at https://angular.dev/license
 */
/**
 * If on a Macintosh with an extended keyboard, the Enter key located in the
 * numeric pad has a different ASCII code.
 */
export declare const MAC_ENTER = 3;
/** The Enter key. */
export declare const ENTER = 13;
/** The Space key. */
export declare const SPACE = 32;
/** Special keycodes used by jsaction for the generic click action. */
export declare const KeyCode: {
    MAC_ENTER: number;
    ENTER: number;
    SPACE: number;
};